+++ src/main/java/com/togedy/togedy_server_v2/global/security/jwt/exception/JwtException.java
@@ -1,4 +1,4 @@
-package com.togedy.togedy_server_v2.global.security.jwt;
+package com.togedy.togedy_server_v2.global.security.jwt.exception;
 
 import com.togedy.togedy_server_v2.global.error.CustomException;
 import com.togedy.togedy_server_v2.global.error.ErrorCode;
