+++ src/main/java/com/togedy/togedy_server_v2/global/security/jwt/exception/JwtInvalidException.java
@@ -1,0 +1,9 @@
+package com.togedy.togedy_server_v2.global.security.jwt.exception;
+
+import com.togedy.togedy_server_v2.global.error.ErrorCode;
+
+public class JwtInvalidException extends JwtException {
+    public JwtInvalidException() {
+        super(ErrorCode.JWT_INVALID);
+    }
+}
