+++ src/main/java/com/togedy/togedy_server_v2/global/security/jwt/exception/JwtMalformedException.java
@@ -1,0 +1,9 @@
+package com.togedy.togedy_server_v2.global.security.jwt.exception;
+
+import com.togedy.togedy_server_v2.global.error.ErrorCode;
+
+public class JwtMalformedException extends JwtException {
+    public JwtMalformedException() {
+        super(ErrorCode.JWT_MALFORMED);
+    }
+}
