+++ src/main/java/com/togedy/togedy_server_v2/global/security/jwt/exception/JwtUnsupportedException.java
@@ -1,0 +1,9 @@
+package com.togedy.togedy_server_v2.global.security.jwt.exception;
+
+import com.togedy.togedy_server_v2.global.error.ErrorCode;
+
+public class JwtUnsupportedException extends JwtException {
+    public JwtUnsupportedException() {
+        super(ErrorCode.JWT_UNSUPPORTED);
+    }
+}
